api/requirements.txt
# requirements.txt
Django==5.0.7
djangorestframework==3.15.2
gunicorn==21.2.0
PyJWT==2.9.0
django-cors-headers==4.4.0
argon2-cffi==23.1.0
reportlab==4.2.2
psycopg[binary]==3.1.18
cryptography>=41.0.0
Pillow==10.4.0
pillow-heif==0.16.0
requests==2.32.3
openpyxl==3.1.5
xlrd==2.0.1
